requirements.txt
# Overseer Python dependencies (shared across receiver, worker, api)
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
redis[hiredis]==5.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
pyotp==2.9.0
qrcode[pil]==7.4.2
httpx==0.27.0
pysnmp>=6.0.0
alembic==1.13.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
anyio[trio]==4.6.0
